Better support guid reconnect - send all windows state
Instead of killing Xorg and restarting itself, send all windows state.
This should make possible to start gui-agent and Xorg during DispVM
savefile preparation and keep them running. So no need to start Xorg
again in time critical phase - DispVM startup.

The improvement we get here, depends on many factors. On some machines
this halves DispVM startup time (especially when disk is slow), on
others the difference is very small.

This is first attempt and not decided if the right one. Alternative is
to enumerate all windows on the X server, instead of using gui-agent
internal state. Not sure which is better, this one is simpler.

Also there is one issue with libvchan API - libvchan_server_init is
blocking. Because of this we can't simply start Xorg and wait for guid
connection, while handling other events and building internal state.
This is one of the reasons why guid is required during dispvm template
preparation. The other one is a need for screen resolution (but it could
be done in some other way).