archlinux: Fix temporarily non working rootless xorg.
The cause seems to be that qrexec run su, which does not create correctly a console logind session. This is simulated by fedora using an unknown mechanism. Probably a pam package such as pam_console.so, which does not exists in archlinux.
One solution (also for portability reasons) would be to use a dedicated pam module instead of su, for which policies varies from one OS to another.