pulse: fix handling record stop event
There was a typo "-=" instead of "==", which was quite fatal - in
pulseaudio 6.0 it causes crash on module unload.
While at it, report as a failure only sending QUBES_PA_SOURCE_START_CMD
to dom0. When sending QUBES_PA_SOURCE_STOP_CMD fails it can simply mean
that dom0 part (pacat-simple-vchan) is no longer running - it can be a
reason why we want to stop recording.