vmside: delay MSG_MFNDUMP until first damage notify event
When window is mapped it should be ready to be displayed (i.e. have its
content filled), so theoretically it is the right place for sending MFNs
list to dom0. But apparently some QT versions first maps the window,
then initialize its content. This cause a race with gui-agent - if
it gets MFNs list before window image is initialized, it would be
actually MFNs list of the root window image. And that wrong list would
not be updated shortly, because gui-agent sends MSG_MFNDUMP only when it
thinks that the image was changed - map event and window size/position
change.

To prevent that race, wait with MSG_MFNDUMP for the first damage
notification after map event. To implement that, add a new field to
window_data struct.
In response to window size change, MSG_MFNDUMP is send immediatelly,
same as before.

qubesos/qubes-issues#1051