Allow qubes_drv reconnection to gui-agent
When the device is disabled (xinput disable), the connection is closed.
Since no state is lost in such case, it's easy to restore connection
(when the device got enabled), so wait for such reconnection.

Permissions of that socket are set in qubes-run-xorg.sh script, so try
to preserve them - otherwise non-root X server will not be able to
reconnect.

Apparently this disable+enable sometimes happens at X server startup.

Fixes QubesOS/qubes-issues#1167