Make qubes-session-env consistent with qrexec-fork-server's environment
`export -p` replaced `env` to fix issues with multiline variables, but
filtering output of `export -p` by grep is still risky. If any of
filtered out variables happens to be multiline, result will be broken.
Unsetting variables in a subshell is safer.

There is no need to unset $_, as `export -p` does not print it,
and no need to restore umask, as it is changed only in subshell.

Currently qubes-session changes umask from 0002 to 0022. I do not know
if it is intentional (maybe author wanted to restore previous value?),
so I am not touching it.

Variables set by gnome-keyring-daemon are appended to qubes-session-env
as declaration, so they will not be exported to environment when
qubes-session-env is sourced by qubes-session.sh. It can be fixed by
dumping environment not at the beginning of qubes-session, but just
before starting qrexec-fork-server.

Additionally qubes-session-env will get more variables (e.g.
QUBES_KEYMAP), but I do not think it is a problem, since they are set in
qrexec-fork-server's environment and passed to user's processes anyway.