window-icon-update: fix crash when all the icons are too big
When all the icons are too big (> 128x128), get_icons will still return
them, but after filtering the list will be empty.
Move the filtering into get_icons function and if that results in an
empty list, raise NoIconError as in other cases.

The strange thing is that this is probably only a side effect of some
incomplete property set/get, as the icon size causing the problem is
odd (like 132x185). But still - even misbehaving application shouldn't
crash icon-sender.

Fixes QubesOS/qubes-issues#1497