Prevent window composition buffers from being moved in RAM
Description by HW42:
The problem is that the pages with the windows content need to be pinned
to a physical page when being mapped with xc_map_foreign_pages. To
archive this the current code calls mlock(). Unfortunately this only
ensures that the page doesn't get swapped out but the kernel can still
move the page.

From [0]:
> A page that has been locked into memory with a call like mlock() is
> required to always be physically present in the system's RAM. [...].
> But there is nothing that requires a locked page to always be present
> in the same place; the kernel is free to move a locked page if the
> need arises.

So when a VM window looks corrupted the VM kernel has moved the page and
it has now a new physical address and in dom0 you now see the new
content of the old physical page.

Unfortunately the patch from [0] didn't got into the kernel, so there
is no easy way to pin a page similar to mlock().

One workaround is to disable memory compactation of unevictable (i.e.
mlocked) pages:

  echo 0 > /proc/sys/vm/compact_unevictable_allowed

This has two downsides:

a) While I didn't observed the effect anymore with this setting I'm not
   sure if this prevents all cases in which the kernel moves pages. (A
   bit grepping suggest that it ignores this flag when allocating
   hugepages but I'm not sure about this.)

b) Given that VMs in Qubes are often low on memory I'm a bit uneasy to
   change memory management options without knowing the side effects.

Properly pinning is not that easy since there seems to be no way to flag
a page as unmovable. The common way seems to be to call
get_user_pages() to increment the reference counter and thereby pin it.
The problem is that then the u2mfn module needs to be changed
significantly to track allocation and deallocation since else the pages
never get freed.

[0]: https://lwn.net/Articles/600502/

Fixes QubesOS/qubes-issues#1028
Thanks HW42 for debugging this.

(cherry picked from commit 9c533813208987ed8e125f6ba6e343a85441a6d3)