vmside: force WindowBorderWidth to zero
X has some (old) features for drawing windows borders through the
X-server. Most X-clients like GTK don't use this feature but some
classic X-clients like xterm do. The problem is that this border is
included in the graphic memory which is shared with dom0. Since it is
ignored in the x/y/width/height values passed to gui-daemon this leads
to croping the window and artifacts due to damage events with wrong
coordinates. This patches simply resets the WindowBorderWidth value to
zero if a X-client sets it to something else. kwin and xfwm seem to do
the same.