pulseaudio: Create a card, a profile and ports
Prior to this commit, Firefox 48 and 49 did not detect a microphone on
Qubes OS AppVMs, even though the pulseaudio module for vchan was
present and running.

Further debugging revealed that Firefox checks for the existence of an
available pulseaudio "port" (similar to a jack) to verify that a
microphone is plugged in.

This commit fills in the blanks by making pulseaudio's vchan module
create a card, a default profile (that is always available), and two
ports (one for the sink, one for the source, both of which are always
available). With this commit, Firefox successfully detects the
microphone on Qubes OS AppVMs.

Reproducer with a -debug build of Firefox:

  env MOZ_LOG="MediaManager:5,GetUserMedia:5,webrtc_trace:65535,timestamp,append,sync" \
    MOZ_LOG_FILE="/home/user/webrtc_log" ./firefox/firefox 'https://meet.jit.si'

After running this command line, and "logging in" to Jitsi Meet,
webrtc_log.* files will contain debugging output. In particular, the
following line is significant:

  D/GetUserMedia Cubeb device 1: type 0x1, state 0x2, name Qubes VCHAN source, id ...

Here, a state of 0x2 indicates a microphone that is available, whereas a
state of 0x0 indicates an unavailable microphone.

Please note that for Jitsi Meet to function properly, the microphone
must be attached to the AppVM running Firefox, using Qubes Manager.

Signed-off-by: M. Vefa Bicakci <m.v.b@runbox.com>