debian: adjust Xorg path for Debian stretch
qubes-gui-agent do not work with suid wrapper for Xorg (because X server
is not "started by a console user"), so it calls Xorg binary directly
(instead of a wrapper). Adjust the wrapper path for actual Xorg binary
path in stretch.

Also, (recent) Xorg when called as normal user will default to logfile
in user home, so no need to provide explicit path in /tmp. Tested on
jessie and stretch.

This solution of QubesOS/qubes-issues#2370 is better than previous,
because it does not depend on some package (xserver-xorg-legacy) not
being installed.
Fixes QubesOS/qubes-issues#2370