qubes-run-xorg: add configurable videoram overhead
When starting X server inside qube, a fixed amount of RAM is allocated
to be used for video buffer. By default, it is enough for the current
desktop (at the time of qube start) and another 1920x1080 desktop, which
may be hotplugged in the future. That is not enough for 4K external
monitor, but we cannot preallocate 4K video buffer for everyone (about
24MiB additional memory per qube).

Now we allocate max(current desktop + overhead, min), where min and
overhead are configurable by qubesdb entries /qubes-gui-videoram-min and
/qubes-gui-videoram-overhead. Default behaviour remains unchanged.