Fix X11 session registration in logind
Most of the logind session registration is done by pam_systemd, but it
needs some extra data to properly register X11 session. Most of it can
be set using environment variables (XDG_SEAT, XDG_VTNR,
XDG_SESSION_CLASS). But at least DISPLAY cannot, it needs to be set with
pam_set_item(3). Unfortunately standard utilities like su or runuser
don't do that, which means we need our own wrapper for that.

Add qubes-gui-runuser, similar to runuser. The main difference is
setting everything needed for logind to recognize the session as
graphical - including PAM_XDISPLAY and use that instead of su. Note that
the qubes-run-xorg script is called as root, so the qubes-gui-runuser
does not need to have suid.

In addition to that, logind require seat to have at least one device
with master-of-seat tag, to consider the seat graphical. VMs have not so
much devices (especially PV or PVH), but each have pcspkr, which
apparently is an input device. Tag it with master-of-seat and attach to
seat0 to satisfy logind requirements.

The last thing is binding X session to a tty. Do that by launching
qubes-gui-agent service connected to tty7, and XDG_VTNR based on control
tty name in qubes-gui-runuser. This require stdin not being closed when
starting qubes-run-xorg.

Fixes QubesOS/qubes-issues#3664