Build package for dom0 too
This is useful if GUI VM and/or Audio VM is not the same as dom0. Then,
frontend is useful in dom0, connecting to a backend running elsewhere.

The audio part is also useful for more complex configurations, like
proxying audio from one VM, through dom0, to sys-usb, to a USB/bluetooth
connected audio output.

QubesOS/qubes-issues#1590