Disable "transparent hugepage" kernel feature
Kernel in Debian has this set to "always" by default, which breaks u2mfn
module. Specifically, it tries to get address of an arbitrary memory
page mapped by X server, and it fails if it's merged into huge page.
u2mfn module is gone in the next Qubes release, so as a temporary
measure revert transparent hugepage to "madvise".

Do that only if it was set to "always". This will prevent overriding
other user choices.

QubesOS/qubes-issues#5212