icon-sender: reconnect on failure
Currently, icon-sender breaks whenever the connection to icon-receiver
fails. This might happen if the user restarts the graphical session (for
example, logs out and logs in again).

Instead, whenever we're disconnected (qrexec-client-vm died), try to
reconnect. This is limited to one try to prevent thrashing.

The solution isn't perfect: after restarting the session, the existing
icons are all gone until you create a new window. This is because we
have no way of telling when icon-receiver is back until we try.