vmside: update cursor position on LeaveNotify
When moving mouse through the window edge, the LeaveNotify event
contains cursor coordinates that are outside of window bounds.
Forward them to local window, otherwise the last event position
the window will receive will be from a MotionNotify still inside
the window.

This fixes an annoying glitch in Firefox: moving mouse cursor up
from a tab title (which touches the top edge of the window) makes
Firefox think we are still hovering over the tab, and displays a
tooltip, which stays there even after you switch to a different
program.