guivm: clean qrexec-service related environment in GUI session
Xephyr server and the GUI (xfce) session running inside is started now
as a qrexec service. Do not let qrexec-related environment variables to
propagate to the actual session, to not confuse some utilities.
This specifically affects qvm-sync-appmenus which behaves differently
when called as a qrexec service. But could affect also
qrexec-fork-server (a service could see gui-session related variables if
for any reason qrexec-fork-server fails to set them to the
actual call values).