Implement VNC GuiVM
We expose a new 'guivm' flavor. In order to enable this remote GuiVM,
one has to enable Qubes services:

- lightdm
- guivm
- guivm-vnc

This is similar to `sys-gui-gpu` setup but here, `guivm-vnc` service
add ExecStartPre and ExecStartPost to lightdm systemd service in order
to configure Xorg server with dummyqbs driver and x11vnc server.