Ensure that no subwindows are created
The GUI daemon won’t accept subwindows for much longer.  Thankfully, the
only ways that a MSG_CREATE with a non-root parent can be sent are
either a managed window being reparented, or a window being created with
an embedder window as parent.  Both are easy enough to check for.