Make qubes-session inherit the systemd user session variables.
The standard environment.d facility is currently not respected in Qubes OS 4.1.  This leads to inability to define environment variables when starting apps via qrexec¹.

This should be fixed.  Here is the fix — it is a fix that will not have any substantial performance impact (it runs *once*), and uses documented APIs to execute on the fix.

¹ Technically there's `/etc/profile.d` but that's meant to be used for **shells**.  There is no guarantee that apps started via qrexec will inherit these variables.