PipeWire audio support for Qubes OS
This adds support for the PipeWire multimedia processing framework.
PipeWire has several advantages over PulseAudio:

- PipeWire supports video as well as audio, which could allow use with
  Qubes Video Companion.

- Not only does PipeWire support its own native API, it also supports
  unmodified JACK and PulseAudio clients.

- PipeWire can be used for low-latency professional audio as well as
  normal deskop use.

- PipeWire modules use a stable API and ABI, and do not require
  including internal PipeWire headers.  On the other hand, PulseAudio
  modules must use the internal (and unstable) PulseCore API.
  Therefore, the PulseAudio module is a maintenance hog, while the
  PipeWire module should need very little maintenance.

- PipeWire is actively maintained by Wim Taymans.  PulseAudio was
  created by Lennart Poettering, but he has since moved on to systemd.

- PipeWire has a permissions system that allows it to be used by
  sandboxed clients.

In the long term, I expect the PipeWire module to replace the PulseAudio
module entirely, but for now, I have allowed selecting between
PulseAudio and PipeWire at runtime.  The default is to use PulseAudio,
but if the `pipewire` service is enabled or if PulseAudio is not
installed, PipeWire will be used instead.  Audio quality appears to be
quite good, at least as far as I can tell.