Implement reconnection and fix bugs
This implements reconnection to a crashed and/or disconnected audio
daemon.  It also works around a nasty bug in PipeWire on Arch Linux that
caused an absurdly incorrect number of bytes to be requested from the
capture stream, and discards any captured sound samples if no capture is
happening.  This is necessary because pacat-simple-vchan can continue to
write audio even after being told to stop, and it uses blocking vchan
writes for this.  Therefore, if data it writes is not read, it will
hang eventually.