Avoid blocking realtime thread
Functions such as libvchan_server_init() and libvchan_close() are not
safe to call on the realtime thread.  They must be called on the main
thread instead.  This requires fairly complex synchronization to ensure
that all functions are called on the correct threads.  spa_loop_invoke
is used to marshal calls between threads.