Avoid potential race conditions and use-after-free
pw_stream_connect() must be called from the main thread, so it must be
invoked via an asynchronous spa_loop_invoke() call.  This means that it
could be called on a stream that was already freed.  The same problem
exists for pw_stream_disconnect(), too.

To avoid these problems, have all functions called via spa_loop_invoke
check a global shutting_down boolean flag and return early if it is set.
The flag is marked as atomic, so it can be checked by both the main
thread and the realtime thread.