Fix segfault on startup
spa_pod_builder_add_object() is variadic, so there is no type checking
done by the C compiler.  The SPA_POD_Int macro expands to _two_
arguments, the first being a string that indicates the type of the
second.  Since this macro was not being used,
spa_pod_builder_add_object() tried to dereference 0x04 (the value of
impl->frame_size) as a pointer, which crashed.

Fix this by using the appropriate macro.