Give input focus to globally active clients
This technically violates the ICCCM, but the alternative is not much
better: these windows *will* have focus in the GUI qube, so not giving
them focus on the agent side will cause the X state to no longer be the
same on both sides.  Worse, such clients might not be expecting to lose
focus without warning.  Explicitly taking focus away from them might
help.

This is based on a similar commit in wlroots.