Support MSG_WINDOW_DUMP_ACK
The GUI agent must not release a grant mapping until it is known that
the GUI daemon has either mapped the grant or will never map it.
Otherwise, the following race is possible:

 1. Agent sends grant ref to daemon.
 2. Agent unmaps grant.
 3. blkfront or netfront allocates same grant number.
 4. Kernel passes grant number to backend VM.
 5. Backend VM maps and uses grant.
 6. Backend VM unmaps grant.
 7. Backend tells frontend that grant has been unmapped.
 8. GUI daemon *finally* maps grant.
 9. Frontend checks that the grant has been unmapped.
10. Frontend thinks this is an attempt to exploit XSA-396 and shuts down
    the device.
11. Qube freezes because one of its block devices stopped working, or
    loses network access because its uplink stopped working.

Fully solving this problem will require kernel changes, but it can be
somewhat worked around in userspace.  Add a new message,
MSG_WINDOW_DUMP_ACK, to inform the agent that a grant will not be mapped
in the future.  It might or might not have been mapped, but this does
not matter.  What matters is that if it has not been mapped so far, it
will not be mapped at all.  The workaround is partial because if the
agent prematurely exits, all grants it owns will be freed by the kernel.
Therefore, the same race is still possible.