Workaround software rendering issues in GTK4
Apply custom CSS to workaround software rendering issues:
 - black frames around popup menus
 - spinner monopolizing the whole "idle loop", preventing actual work to
   be done (for example in nautilus)

This is heavily based on work by @sityvb and @noskb.

The spinner disabler is done in two parts:
 - animation-play-state disables actual spinning (this part frees up the
   idle loop)
 - opacity hides the now frozen spinner to avoid confusion (at least a
   bit); it could be 'display: none', but that (at least in nautilus) is
   programatically overridden

The patch is applied on user login and flag file is created to not
reapply it needlessly.

Fixes QubesOS/qubes-issues#8081
Fixes QubesOS/qubes-issues#7921
Fixes QubesOS/qubes-issues#8115