Do not rely on pw_context_get_support() on modern PipeWire
pw_context_get_support() is guaranteed to return a non-NULL pointer that
can be used with spa_support_find(), but does not guarantee that any
particular interface is present in the returned support objects.  The
list of interfaces has changed in the past and might change again.
Instead, use the public functions in the PipeWire API, which are
more likely to remain stable in future releases of PipeWire.