Prevent $DISPLAY from being overridden in SSH (for X11 forwarding)
When trying to ssh into a qubes vm, `/etc/profile.d/qubes-gui.sh` will
improperly set `$DISPLAY` to be :0. When X11 forwarding is used, `$DISPLAY`
is automatically set to ensure that the relevant windows get forwarded
to the ssh client. However, since qubes-gui.sh is called after a
successful login, the correct value for `$DISPLAY` will be overridden.

This commit fixes the issue by adding a test in qubes-gui.sh to check if
we are currently in an ssh session (by checking if `$SSH_CLIENT` or
`$SSH_TTY` are set).

Note: I do not know .csh scripts and have not tested the proposed change
to qubes-gui.csh, yet.