Disable auto-add GPU by default
The dummyqbs driver makes use of GPU render engine itself, and Xorg
trying to use that as separate device would conflict with it. But if
somebody want to use that anyway (to use a monitor connected to the
secondary GPU), it can be re-enabled manually using xorg.conf.d snippet.
But better, use sys-gui-gpu for that instead (or standard lightdm/sddm
etc instead of qubes-gui-agent).

This is yet another part of easing GPU passthrough for rendering
acceleration.