Do not allocate grants for framebuffer by default
If glamor is not used (which is default), FBBasePriv seems to be unused.
GUI agent doesn't share the root window with GUI daemon. If glamor is
used, FBBasePriv probably unused in practice too - the GUI agent doesn't
support sharing part of the pixmap anyway, but I'm not 100% sure about
it (see qubes_create_screen_resources() - the "pixmap" of the root
window is passed to glamor, and it isn't clear if it wouldn't end up
used in some window that is handled by the GUI agent).

Based on this observation, use normal memory for the framebuffer. The
main benefit is not having the framebuffer mlock()-ed, which helps
especially with small VMs. But also, allocating it as normal userspace
memory gives the kernel more flexibility in finding memory for it (it
doesn't need to be physically continuous, etc).
Due to the glamor case, leave the grant-based FBBase allocation in the
code, but disabled by default. And leave a #define to re-enable it, with
appropriate comment.

Fixes QubesOS/qubes-issues#9992