Increase screen pixmap size beyond initial video RAM if needed
If FBBase is not backed by grant tables, it's now possible to reallocate
it. This is especially useful when new external display is connected and
the overall resolution is increased.

Note the realloc() call may change FBBase address. In practice, FBBase
is given to fbScreenInit, and it does save it at some point
(fbScreenInit->fbFinishScreenInit->miScreenInit->miScreenDevPrivateInit),
but then it's used only to attach it to the screen pixmap, which is
updated via ModifyPixmapHeader() call few lines below anyway.

This change adjusts also pScrn->videoRam. But fortunately X server seems
to use it only at startup, and only to validate initial modes list.

Fixes QubesOS/qubes-issues#7448