Don't try to send grants for not realized windows
We rely on composite redirect mode of the X server to get per window
pixmaps. Those are setup/teared-down in compRealizeWindow/
compUnrealizeWindow (via compCheckRedirect).

So not realized windows don't have a per window pixmap. Sending grant
refs for them was always broken since we didn't send the offset into the
screen pixmap in those cases. But with the recent change to not allocate
grant refs for the screen pixmap this leads to a noticable error
message.

So don't try to send grant refs for not realized windows. This means
that the configure before mapping will not contiain grant refs.

We rely here very much on implementation details of the X server. In
particular we require that a damage event is generated after the windows
is realized. Note that the map event is generated before realization.
Additionally we require that no damage event is delivered between the
map notification the window being realized.

Based on experimentation and code reading that seems to hold (what
exactly triggers the first damage event after map depends on background
settings and cursor position). This matches also that we so far haven't
had problems even though this bug has been present for a long time.