Don't try to send grants for not realized windows
We rely on composite redirect mode of the X server to get per window
pixmaps. Those are setup/teared-down in compRealizeWindow/
compUnrealizeWindow (via compCheckRedirect).

So not realized windows don't have a per window pixmap. Sending grant
refs for them was always broken since we didn't send the offset into the
screen pixmap in those cases. But with the recent change to not allocate
grant refs for the screen pixmap this leads to a noticable error
message.

So don't try to send grant refs for not realized windows. This means
that the configure before mapping will not contiain grant refs. But when
we map the window we will get a damage event because of the new pixmap
compRealizeWindow has allocated and send them then. So this should be
fine.