Switch to tty7 instead of calling loginctl activate
loginctl activate was a workaround to activate the session, but the
proper solution is to activate tty on which the session is running.
More importantly, this approach does not relies on polkit to allow
loginctl call.

Do this by calling VT_ACTIVATE ioctl on stdin - which is a tty.

Fixes QubesOS/qubes-issues#10422